pyproject.toml
[project]
name = "mfsg-camera"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
]
